streamlit_app/requirements.txt
# Streamlit Dashboard Requirements
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
